weather_data_downloader/requirements.txt
meteostat~=1.6.7
omegaconf~=2.3.0
hydra-core~=1.3.2